refactor: split the cli entry into per-command modules
