+++ packages/cli/src/cli/options.ts
@@ -1,0 +1,4 @@
+/** Commander value collector for repeatable options. */
+export function collect(value: string, previous: string[]): string[] {
+  return previous.concat([value])
+}
